+++ src/components/notes/notebook-view.tsx
